Rename core model implementation serializable > pojo
