+++ model-core-test/src/main/java/net/spals/appbuilder/model/core/pojo/BuilderClass.java
@@ -1,4 +1,4 @@
-package net.spals.appbuilder.model.core.serializable;
+package net.spals.appbuilder.model.core.pojo;
 
 import org.inferred.freebuilder.FreeBuilder;
